requirements.txt
Django>=5.1,<6.0
psycopg[binary]>=3.2
python-dotenv>=1.0
pandas>=2.2
